Refac : [002] Improve authentication

+++ server/controllers/delete.js
@@ -2,18 +2,25 @@ const productModel = require("../models/products");
 
 function deleteHandler(req, res) {
   try {
-    productModel
-      .destroy({ where: { uid: req.params.uid } })
-      .then(() => {
-        res.status(200).json({ status: "200: OK", message: "Item deleted" });
-      })
-      .catch((err) => {
-        console.log(`Error: controllers/delete.js \n${err}`);
-        res.status(400).json({
-          status: "400: Bad request",
-          message: "Unable to delete item due to client error",
+    if (req.user.userType === "seller") {
+      productModel
+        .destroy({ where: { uid: req.params.uid } })
+        .then(() => {
+          res.status(200).json({ status: "200: OK", message: "Item deleted" });
+        })
+        .catch((err) => {
+          console.log(`Error: controllers/delete.js \n${err}`);
+          res.status(400).json({
+            status: "400: Bad request",
+            message: "Unable to delete item due to client error",
+          });
         });
+    } else {
+      res.status(400).json({
+        status: "400: Bad request",
+        message: "Unauthorized seller",
       });
+    }
   } catch (err) {
     console.log(`Error: controllers/delete.js \n${err}`);
     res.status(500).json({
